NETWORK_UTILITY/requirements.txt
flask
Flask-Compress
Flask-Cors
PyYAML
netifaces
netaddr